+++ app/Providers/Filament/AdminPanelProvider.php
@@ -59,6 +59,7 @@ class AdminPanelProvider extends PanelProvider
             ->navigationGroups([
                 'PayPal',
                 'pretix',
+                'Bank',
                 'Events & Kunden',
                 'Transaktionen',
                 'Berichte',